Refactor OFB class to use boolean arrays for IV
Updated the `IV` variable to be a `bool[]` instead of a `string`, and modified the constructor to initialize it from a string input. Changed the `Encrypt` method to accept `bool[][]` messages and return encrypted boolean arrays. The `PermutationCipher` and `XOR` methods were also refactored to work with boolean arrays, simplifying their implementations.